+++ src/Codec/Xlsx/Parser/Internal.hs
@@ -14,6 +14,7 @@ module Codec.Xlsx.Parser.Internal
     , readFailure
     , invalidText
     , defaultReadFailure
+    , boolean
     , decimal
     , rational
     ) where
@@ -23,7 +24,7 @@ import           Data.Text (Text)
 import qualified Data.Text as T
 import qualified Data.Text.Read as T
 import           Data.Typeable (Typeable)
-import           Data.XML.Types
+import           Text.XML
 import           Text.XML.Cursor
 
 #if !MIN_VERSION_base(4,8,0)
@@ -107,9 +108,15 @@ n x = Name
 decimal :: Monad m => Text -> m Int
 decimal t = case T.decimal t of
   Right (d, _) -> return d
-  _ -> fail "invalid decimal"
+  _ -> fail $ "invalid decimal" ++ show t
 
 rational :: Monad m => Text -> m Double
 rational t = case T.rational t of
   Right (r, _) -> return r
-  _ -> fail "invalid rational"
+  _ -> fail $ "invalid rational: " ++ show t
+
+boolean :: Monad m => Text -> m Bool
+boolean t = case T.strip t of
+    "true"  -> return True
+    "false" -> return False
+    _       -> fail $ "invalid boolean: " ++ show t